requirements.txt
Flask==2.3.3
google-generativeai==0.2.0
requests==2.31.0
